requirements.txt
flask
flask-limiter
flask-caching
redis
numpy
scikit-learn
python-dotenv
gunicorn
fastapi
aiohttp
beautifulsoup4
lxml
torch
sentry-sdk
pydantic
pydantic-settings
structlog
elasticsearch
transformers
python-jose[cryptography]
uvicorn